+++ config/routes.rb
@@ -2,5 +2,7 @@
 
 Docit::Engine.routes.draw do
   root to: "ui#index"
+  get "swagger", to: "ui#swagger"
+  get "scalar", to: "ui#scalar"
   get "spec", to: "ui#spec", defaults: { format: :json }
 end